renaming job post context file
